# Optimize Class#new and aliased methods