@meh 2017-02-25
Just a heads up, we have a PR in-flight (#1637) that touches that `new_body` stuff.