@meh 2017-02-25
I'm okay with `Opal.send` all the things, especially if it makes a difference wrt performance.

I feel less okay with the current *hand compiled* `Class.new` we have here, moving it onto the code generator might be a better idea as you mention, since it's likely to improve performance across the board instead of just for that specific function.

Could you maybe make a diff or point out the differences between the code that's currently generated and the *hand compiled* version?