@meh 2017-02-25
Also we should probably merge the `new_body` changes separately so the two pull requests don't have to stomp eachother's toes, since there's clearly a performance win there.