@elia 2017-02-26
@jgaskins great stuff, I was looking at ways to optimize the alias body wrapper and was even ready to go back to using the original body (see 2c4827313) at the cost of being less compliant.

Looking at `Class.new` here's the diff with the generated code:

```rb
class Class
  def new(*args, &block)
    %x{
      var obj = #{allocate};

      obj.$initialize.$$p = block;
      obj.$initialize.apply(obj, args);
      return obj;
    }
  end
end
```

```js
  /* Generated by Opal 0.11.0.dev */
  (function(Opal) {
    var self = Opal.top, $scope = Opal, $scopes = [Opal], nil = Opal.nil, $breaker = Opal.breaker, $slice = Opal.slice, $klass = Opal.klass;

    Opal.add_stubs(['$allocate']);
    return (function($base, $super, $visibility_scopes) {
      function $Class(){};
      var self = $Class = $klass($base, $super, 'Class', $Class);

      var def = self.$$proto, $scope = self.$$scope, $scopes = $visibility_scopes.slice().concat($scope), TMP_Class_new_1;

      return (Opal.defn(self, '$new', TMP_Class_new_1 = function($a_rest) {
        var self = this, args, $iter = TMP_Class_new_1.$$p, block = $iter || nil;

        var $args_len = arguments.length, $rest_len = $args_len - 0;
        if ($rest_len < 0) { $rest_len = 0; }
        args = new Array($rest_len);
        for (var $arg_idx = 0; $arg_idx < $args_len; $arg_idx++) {
          args[$arg_idx - 0] = arguments[$arg_idx];
        }
        TMP_Class_new_1.$$p = null;
      
        var obj = self.$allocate();

        obj.$initialize.$$p = block;
        obj.$initialize.apply(obj, args);
        return obj;
    
      }, TMP_Class_new_1.$$arity = -1, TMP_Class_new_1.$$parameters = [['rest', 'args'], ['block', 'block']]), nil) && 'new'
    })($scope.base, null, $scopes)
  })(Opal);
```

_generated with `pbpaste | be bin/opal -cEOA | pbcopy`_


Except the fact that `Class$new` looks very good seems that the main diff is `Opal.send`, so I'd keep just that change there.