@jgaskins 2017-02-26
@elia Using the original body was something I added in #1204 because I didn't understand why it was wrapped and, based on the comments, nobody else was sure, either. :-) I was surprised to see the wrapper was added back in, but since you mentioned it being less compliant, it sounds like you found the purpose for it?

The other (minor) difference besides `Opal.send` between the current wrapper and the one in this PR is unconditional vs conditional clearing of `$$p`. I'm not sure if that's what was causing the de-opt, but it's possible.