@elia 2017-03-07
> it sounds like you found the purpose for it?

Yes, basically using a different body helps with `Method` and `UnboundMethod` returning the correct owner. Also I think it's useful when dealing with super. A way to clone a function would have helped a lot with performance here.

> The other (minor) difference besides Opal.send between the current wrapper and the one in this PR is unconditional vs conditional clearing of $$p. I'm not sure if that's what was causing the de-opt, but it's possible.

Sure sorry, what I really meant is to keep wrap the relevant changes in the usual `def self.` wrap.

@jgaskins ping me when done and I'll merge.