@elia 2017-03-11
@jgaskins I added the required changes and a further alias optimization in:

- d4676e2f67 (Avoid Chrome deoptim by conditionally reass. block)
- 49026aab3d (Improve docs and rename method wrapper in .alias())
- fbe1e58b01 (Avoid long wrapper chains in Opal.alias())